HPERNIT.FOR -- reverse ammonia and nitrate labels for hbn file HRCHSED.FOR -- change conversion factor for sediment to be more precise
BTW, the internal mass units in SEDTRN are mg*ft3/liter.  The HRCHSED.FOR bug was described as follows:

On 1985/10/ 2 20, PERLND 101 produces 1.801E-07 tons/ac of sedment, which you can see in sed_per.out.
Since there is only 1 acre of PERLND 101, we'd expect the same amt of sediment to reach RCHRES 1 at that time.  But RCHRES 1 receives an inflow of 1.798E-07 tons of sedment, as shown in sed_riv.out. The issue related to converting to internal units, and the precision of that conversion.